Fixed keyboard_control node
