+++ src/surface/flight_control/launch/keyboard_control_launch.py
@@ -1,5 +1,5 @@
-from launch.actions import GroupAction
 from launch.launch_description import LaunchDescription
+from launch.substitutions.launch_configuration import LaunchConfiguration
 from launch_ros.actions import Node, PushRosNamespace
 
 
@@ -7,32 +7,13 @@ def generate_launch_description() -> LaunchDescription:
     keyboard_control_node = Node(
         package='flight_control',
         executable='keyboard_control_node',
+        parameters=[
+            {'controller_mode': LaunchConfiguration('controller_mode', default=0)},
+            {'controller_profile': LaunchConfiguration('controller_profile', default=0)},
+        ],
+        remappings=[('/surface/mavros/manual_control/send', '/tether/mavros/manual_control/send')],
         emulate_tty=True,
         output='screen',
     )
 
-    control_inverter_node = Node(
-        package='flight_control',
-        executable='control_inverter_node',
-        emulate_tty=True,
-        output='screen',
-    )
-
-    multiplexer_node = Node(
-        package='flight_control',
-        executable='multiplexer',
-        remappings=[('/surface/mavros/rc/override', '/tether/mavros/rc/override')],
-        emulate_tty=True,
-        output='screen',
-    )
-
-    namespace_launch: GroupAction = GroupAction(
-        actions=[
-            PushRosNamespace('surface'),
-            keyboard_control_node,
-            control_inverter_node,
-            multiplexer_node,
-        ]
-    )
-
-    return LaunchDescription([namespace_launch])
+    return LaunchDescription([keyboard_control_node])
